tests/wpt/tests/css/css-display/reading-flow/tentative/reading-flow-invalid.html

<!DOCTYPE html>
<title>CSS Display: parsing reading-flow with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-display-4/#reading-flow">
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_invalid_value('reading-flow', 'auto');
test_invalid_value('reading-flow', 'none');
test_invalid_value('reading-flow', 'legacy');
test_invalid_value('reading-flow', 'normal auto');
test_invalid_value('reading-flow', 'flex-visual flex-flow');
test_invalid_value('reading-flow', '100%');
test_invalid_value('reading-flow', '10px');
</script>
</body>
